+++ src/main/java/dev/orne/beans/converters/IdentityBeanConverter.java
@@ -1,0 +1,279 @@
+package dev.orne.beans.converters;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import javax.validation.constraints.NotNull;
+
+import org.apache.commons.beanutils.ConversionException;
+import org.apache.commons.beanutils.Converter;
+import org.apache.commons.beanutils.converters.AbstractConverter;
+
+import dev.orne.beans.BaseIdentityBean;
+import dev.orne.beans.Identity;
+import dev.orne.beans.IdentityBean;
+import dev.orne.beans.WritableIdentityBean;
+
+/**
+ * Implementation of {@code Converter} that converts {@code IdentityBean}
+ * instances to {@code String} and {@code WritableIdentityBean} from
+ * {@code String} using the identity token of the bean as {@code String}
+ * representation.
+ * <p>
+ * Uses a nested {@code Converter}, of type {@code IdentityConverter} by
+ * default, to convert bean identities.
+ * <p>
+ * When converting from {@code String} instances of {@code TokenIdentity}
+ * are produced an populated on empty instances of the bean, producing valid
+ * {@code ValidBeanIdentity} instances.
+ * 
+ * @author <a href="mailto:wamphiry@orne.dev">(w) Iker Hernaez</a>
+ * @version 1.0, 2020-09
+ * @since 0.3
+ * @see IdentityBean
+ * @see IdentityConverter
+ */
+public class IdentityBeanConverter
+extends AbstractConverter {
+
+    /** Message error when target type is not WritableIdentityBean. */
+    private static final String NOT_WRITABLE_IBEAN_ERROR =
+            "Can't convert value '%s' to type %s. " +
+            "Target type does not implement %s";
+    /** Message error when cannot instantiate target type. */
+    private static final String NEW_INSTANCE_ERROR =
+            "Can't convert value '%s' to type %s. " +
+            "Cannot create new instances of target type";
+
+    /** The default target type. */
+    private final Class<? extends WritableIdentityBean> defaultType;
+    /** The converter for identities. */
+    private final Converter identityConverter;
+
+    /**
+     * Creates a new instance that throws a {@code ConversionException} if an
+     * error occurs and uses default instance of {@code IdentityConverter} for
+     * identity conversions.
+     */
+    public IdentityBeanConverter() {
+        super();
+        this.defaultType = BaseIdentityBean.class;
+        this.identityConverter = new IdentityConverter();
+    }
+
+    /**
+     * Creates a new instance that returns a default value if an error occurs 
+     * and uses default instance of {@code IdentityConverter} for identity
+     * conversions.
+     * 
+     * @param defaultValue The default value to be returned if the value to be
+     * converted is missing or an error occurs converting the value
+     */
+    public IdentityBeanConverter(
+            final @Nullable IdentityBean defaultValue) {
+        super(defaultValue);
+        this.defaultType = BaseIdentityBean.class;
+        this.identityConverter = new IdentityConverter();
+    }
+
+    /**
+     * Creates a new instance that throws a {@code ConversionException} if an
+     * error occurs.
+     * 
+     * @param identityConverter The converter to use for identity conversions
+     */
+    public IdentityBeanConverter(
+            final @NotNull Converter identityConverter) {
+        super();
+        this.defaultType = BaseIdentityBean.class;
+        this.identityConverter = identityConverter;
+    }
+
+    /**
+     * Creates a new instance that returns a default value if an error occurs.
+     * 
+     * @param identityConverter The converter to use for identity conversions
+     * @param defaultValue The default value to be returned if the value to be
+     * converted is missing or an error occurs converting the value
+     */
+    public IdentityBeanConverter(
+            final @NotNull Converter identityConverter,
+            final @Nullable WritableIdentityBean defaultValue) {
+        super(defaultValue);
+        this.defaultType = BaseIdentityBean.class;
+        this.identityConverter = identityConverter;
+    }
+
+    /**
+     * Creates a new instance that throws a {@code ConversionException} if an
+     * error occurs and uses default instance of {@code IdentityConverter} for
+     * identity conversions.
+     */
+    public IdentityBeanConverter(
+            final @NotNull Class<? extends WritableIdentityBean> defaultType) {
+        super();
+        this.defaultType = defaultType;
+        this.identityConverter = new IdentityConverter();
+    }
+
+    /**
+     * Creates a new instance that returns a default value if an error occurs 
+     * and uses default instance of {@code IdentityConverter} for identity
+     * conversions.
+     * 
+     * @param defaultValue The default value to be returned if the value to be
+     * converted is missing or an error occurs converting the value
+     */
+    public <T extends WritableIdentityBean> IdentityBeanConverter(
+            final @NotNull Class<T> defaultType,
+            final @Nullable T defaultValue) {
+        super(defaultValue);
+        this.defaultType = defaultType;
+        this.identityConverter = new IdentityConverter();
+    }
+
+    /**
+     * Creates a new instance that throws a {@code ConversionException} if an
+     * error occurs.
+     * 
+     * @param identityConverter The converter to use for identity conversions
+     */
+    public IdentityBeanConverter(
+            final @NotNull Class<? extends WritableIdentityBean> defaultType,
+            final @NotNull Converter identityConverter) {
+        super();
+        this.defaultType = defaultType;
+        this.identityConverter = identityConverter;
+    }
+
+    /**
+     * Creates a new instance that returns a default value if an error occurs.
+     * 
+     * @param identityConverter The converter to use for identity conversions
+     * @param defaultValue The default value to be returned if the value to be
+     * converted is missing or an error occurs converting the value
+     */
+    public <T extends WritableIdentityBean> IdentityBeanConverter(
+            final @NotNull Class<T> defaultType,
+            final @NotNull Converter identityConverter,
+            final @Nullable T defaultValue) {
+        super(defaultValue);
+        this.defaultType = defaultType;
+        this.identityConverter = identityConverter;
+    }
+
+    /**
+     * Returns the converter for identities.
+     * 
+     * @return The converter for identities
+     */
+    protected Converter getIdentityConverter() {
+        return this.identityConverter;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    @Nonnull
+    protected Class<?> getDefaultType() {
+        return this.defaultType == null ? WritableIdentityBean.class : this.defaultType;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected <T> T convertToType(
+            @Nonnull
+            final Class<T> type,
+            @Nonnull
+            final Object value) {
+        if (IdentityBean.class.isAssignableFrom(type)) {
+            if (type.isInstance(value)) {
+                return type.cast(value);
+            } else if (WritableIdentityBean.class.isAssignableFrom(type)) {
+                final Identity identity = convertIdentity(value);
+                if (identity == null) {
+                    return null;
+                }
+                final T bean = createInstance(type, value);
+                ((WritableIdentityBean) bean).setIdentity(identity);
+                return bean;
+            } else {
+                throw new ConversionException(
+                        String.format(
+                            NOT_WRITABLE_IBEAN_ERROR,
+                            value,
+                            type,
+                            WritableIdentityBean.class));
+            }
+        } else {
+            throw conversionException(type, value);
+        }
+    }
+
+    /**
+     * Tries to create a new instance of the target type.
+     * 
+     * @param <T> The type of bean to create
+     * @param type The type of bean to create
+     * @param value The input value to be converted. Used only when an error
+     * occurs
+     * @return The newly created instance
+     * @throws ConversionException If the new instance cannot be created
+     */
+    protected <T> T createInstance(
+            final Class<T> type,
+            final @Nonnull Object value) {
+        try {
+            return type.newInstance();
+        } catch (final InstantiationException | IllegalAccessException e) {
+            throw new ConversionException(
+                    String.format(
+                        NEW_INSTANCE_ERROR,
+                        value,
+                        type),
+                    e);
+        }
+    }
+
+    /**
+     * Converts the value to a {@code Identity} instance.
+     * 
+     * @param value The input value to be converted
+     * @return The converted value
+     * @throws ConversionException If an error occurs converting the value
+     */
+    protected Identity convertIdentity(
+            final @Nonnull Object value) {
+        final Identity identity;
+        if (value instanceof Identity) {
+            identity = (Identity) value;
+        } else {
+            identity = this.identityConverter.convert(
+                    Identity.class,
+                    value);
+        }
+        return identity;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    protected String convertToString(
+            @Nonnull
+            final Object value)
+    throws Throwable {
+        if (value instanceof IdentityBean) {
+            return this.identityConverter.convert(
+                    String.class,
+                    ((IdentityBean) value).getIdentity());
+        } else if (value instanceof String) {
+            return value.toString();
+        } else {
+            throw conversionException(String.class, value);
+        }
+    }
+
+}